requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.7.2
    # via django
django==5.0.1
    # via
    #   -r requirements.in
    #   django-admin-honeypot-updated-2021
    #   django-axes
    #   django-debug-toolbar
    #   django-filter
    #   djangorestframework
django-admin-honeypot-updated-2021==1.2.0
    # via -r requirements.in
django-axes==6.3.0
    # via -r requirements.in
django-debug-toolbar==4.2.0
    # via -r requirements.in
django-environ==0.11.2
    # via -r requirements.in
django-filter==23.5
    # via -r requirements.in
django-ipware==6.0.3
    # via django-admin-honeypot-updated-2021
djangorestframework==3.14.0
    # via -r requirements.in
markdown==3.5.1
    # via -r requirements.in
psycopg2==2.9.9
    # via -r requirements.in
python-ipware==2.0.1
    # via django-ipware
pytz==2023.3.post1
    # via djangorestframework
sqlparse==0.4.4
    # via
    #   django
    #   django-debug-toolbar

# The following packages are considered to be unsafe in a requirements file:
# setuptools
